simplify rhai-tpl and add corrosion-specific logic to corro-tpl, add --once flag to corrosion template
