+++ crates/corro-tpl/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "corro-tpl"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+rhai = { workspace = true }
+rhai-tpl = { path = "../rhai-tpl" }
+corro-client = { path = "../corro-client" }
+hostname = { workspace = true }
+tracing = { workspace = true }
+tokio = { workspace = true }
+hyper = { workspace = true }
+tokio-util = { workspace = true }
+bytes = { workspace = true }
+corro-types = { path = "../corro-types" }
+serde_json = { workspace = true }
+serde = { workspace = true }
+compact_str = { workspace = true }
+futures = { workspace = true }
+tripwire = { path = "../tripwire" }
+hex = { workspace = true }
+indexmap = { workspace = true }
+parking_lot = { workspace = true }
+uuid = { workspace = true }
+csv = { version = "1.2.2" }
+thiserror = { workspace = true } 
+enquote = { workspace = true }
+
+[dev-dependencies]
+tracing-subscriber = { workspace = true }
+tempfile = { workspace = true }
+corro-tests = { path = "../corro-tests" }
