+++ crates/rhai-tpl/src/lib.rs
@@ -1,36 +1,16 @@
-use std::collections::HashMap;
 use std::io::BufWriter;
-use std::io::Seek;
 use std::io::Write;
+use std::ops::Deref;
+use std::ops::DerefMut;
 use std::sync::Arc;
-use std::time::Duration;
 
-use bytes::BytesMut;
-use compact_str::CompactString;
-use compact_str::ToCompactString;
-use corro_client::CorrosionApiClient;
-use corro_types::api::QueryEvent;
-use corro_types::api::Statement;
-use corro_types::change::SqliteValue;
-use futures::StreamExt;
-use indexmap::IndexMap;
 use logos::Logos;
 use parking_lot::RwLock;
 use rhai::Dynamic;
+use rhai::EvalAltResult;
 use rhai::Scope;
-use rhai::{EvalAltResult, Map};
-use serde::ser::{SerializeSeq, Serializer};
-use serde_json::ser::Formatter;
-use tokio::sync::OnceCell;
-use tokio::sync::{mpsc, RwLock as TokioRwLock};
-use tokio::time::sleep;
-use tokio_util::codec::{Decoder, LinesCodec};
-use tokio_util::sync::CancellationToken;
-use tracing::debug;
 use tracing::error;
 use tracing::trace;
-use tracing::warn;
-use uuid::Uuid;
 
 #[derive(Debug, PartialEq, Clone, Copy, Default, thiserror::Error)]
 #[error("parse error")]
@@ -59,347 +39,27 @@ enum Closing {
 }
 
 #[derive(Clone)]
-struct QueryResponse {
-    query: Arc<TokioRwLock<QueryHandle>>,
+pub struct TemplateWriter<S> {
+    writer: Arc<RwLock<BufWriter<Box<dyn Write + Send + Sync + 'static>>>>,
+    pub state: S,
 }
 
-impl QueryResponse {
-    fn to_json(&mut self) -> SqlToJson {
-        SqlToJson {
-            json_output: JsonOutput::default(),
-            res: self.clone(),
+impl<S> Write for TemplateWriter<S> {
+    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
+        self.writer.write().write(buf)
+    }
+
+    fn flush(&mut self) -> std::io::Result<()> {
+        self.writer.write().flush()
+    }
+}
+
+impl<S> TemplateWriter<S> {
+    pub fn new<W: Write + Send + Sync + 'static>(w: W, state: S) -> Self {
+        Self {
+            writer: Arc::new(RwLock::new(BufWriter::new(Box::new(w)))),
+            state,
         }
-    }
-
-    fn to_json_w_options(&mut self, options: Map) -> SqlToJson {
-        let json_output = {
-            let pretty = options
-                .get("pretty")
-                .and_then(|d| d.as_bool().ok())
-                .unwrap_or(false);
-            let row_values_as_array = options
-                .get("row_values_as_array")
-                .and_then(|d| d.as_bool().ok())
-                .unwrap_or(false);
-
-            JsonOutput {
-                pretty,
-                row_values_as_array,
-            }
-        };
-
-        SqlToJson {
-            json_output,
-            res: self.clone(),
-        }
-    }
-
-    fn to_csv(&mut self) -> SqlToCsv {
-        SqlToCsv { res: self.clone() }
-    }
-}
-
-#[derive(Clone)]
-pub struct SqlToCsv {
-    res: QueryResponse,
-}
-
-#[derive(Clone)]
-pub struct SqlToJson {
-    json_output: JsonOutput,
-    res: QueryResponse,
-}
-
-impl IntoIterator for QueryResponse {
-    type Item = Result<Row, Box<EvalAltResult>>;
-
-    type IntoIter = QueryResponseIter;
-
-    fn into_iter(self) -> Self::IntoIter {
-        QueryResponseIter {
-            query: self.query.clone(),
-            body: OnceCell::new(),
-            codec: LinesCodec::new(),
-            buf: BytesMut::new(),
-            handle: tokio::runtime::Handle::current(),
-            done: false,
-            columns: None,
-        }
-    }
-}
-
-#[derive(Clone)]
-struct Row {
-    #[allow(dead_code)]
-    id: i64,
-    columns: Arc<Vec<CompactString>>,
-    cells: Vec<SqliteValue>,
-}
-
-impl IntoIterator for Row {
-    type Item = Cell;
-
-    type IntoIter = RowIter;
-
-    fn into_iter(self) -> Self::IntoIter {
-        RowIter {
-            pos: 0,
-            row: Arc::new(self),
-        }
-    }
-}
-
-struct RowIter {
-    row: Arc<Row>,
-    pos: usize,
-}
-
-impl Iterator for RowIter {
-    type Item = Cell;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.row.cells.len() > self.pos {
-            let cell = Cell {
-                index: self.pos,
-                row: self.row.clone(),
-            };
-            self.pos += 1;
-            Some(cell)
-        } else {
-            None
-        }
-    }
-}
-
-#[derive(Clone)]
-struct Cell {
-    index: usize,
-    row: Arc<Row>,
-}
-
-#[derive(Clone)]
-struct SqliteValueWrap(SqliteValue);
-
-impl SqliteValueWrap {
-    fn to_json(&mut self) -> String {
-        match &self.0 {
-            SqliteValue::Null => "null".into(),
-            SqliteValue::Integer(i) => i.to_string(),
-            SqliteValue::Real(f) => f.to_string(),
-            SqliteValue::Text(t) => enquote::enquote('"', &t),
-            SqliteValue::Blob(b) => hex::encode(b.as_slice()),
-        }
-    }
-}
-
-impl Cell {
-    pub fn name(&mut self) -> Result<String, Box<EvalAltResult>> {
-        Ok(self
-            .row
-            .columns
-            .get(self.index)
-            .ok_or_else(|| Box::new(EvalAltResult::from("cell does not exist")))?
-            .to_string())
-    }
-
-    pub fn value(&mut self) -> Result<SqliteValueWrap, Box<EvalAltResult>> {
-        Ok(SqliteValueWrap(
-            self.row
-                .cells
-                .get(self.index)
-                .ok_or_else(|| Box::new(EvalAltResult::from("cell does not exist")))?
-                .clone(),
-        ))
-    }
-}
-
-struct QueryResponseIter {
-    query: Arc<TokioRwLock<QueryHandle>>,
-    body: OnceCell<hyper::Body>,
-    buf: BytesMut,
-    codec: LinesCodec,
-    handle: tokio::runtime::Handle,
-    done: bool,
-    columns: Option<Arc<Vec<CompactString>>>,
-}
-
-impl QueryResponseIter {
-    pub async fn body(&mut self) -> Result<&mut hyper::Body, Box<EvalAltResult>> {
-        self.body
-            .get_or_try_init(|| async {
-                match self.query.write().await.body().await {
-                    Ok((body, _)) => Ok(body),
-                    Err(e) => Err(Box::new(EvalAltResult::from(e.to_string()))),
-                }
-            })
-            .await?;
-        self.body.get_mut().ok_or_else(|| {
-            Box::new(EvalAltResult::from(
-                "unexpected error, body is gone from OnceCell",
-            ))
-        })
-    }
-
-    // pub async fn columns(&mut self) -> Result<Arc<Vec<CompactString>>, Box<EvalAltResult>> {
-    //     {
-    //         if let Some(cols) = self.columns.clone() {
-    //             return Ok(cols);
-    //         }
-    //     }
-
-    //     if let Some(Err(e)) = self.recv().await {
-    //         return Err(e);
-    //     }
-    //     if let Some(cols) = self.columns.clone() {
-    //         return Ok(cols);
-    //     }
-    //     return Err(Box::new(EvalAltResult::from("could not get columns")));
-    // }
-
-    pub async fn read_bytes(&mut self) -> Result<usize, Box<EvalAltResult>> {
-        let bytes_res = {
-            let body = match self.body().await {
-                Ok(body) => body,
-                Err(e) => {
-                    self.done = true;
-                    return Err(e);
-                }
-            };
-            body.next().await
-        };
-        let len = match bytes_res {
-            Some(Ok(b)) => {
-                // debug!("read {} bytes", b.len());
-                self.buf.extend_from_slice(&b);
-                b.len()
-            }
-            Some(Err(e)) => {
-                self.done = true;
-                return Err(Box::new(EvalAltResult::from(e.to_string())));
-            }
-            None => {
-                self.done = true;
-                0
-            }
-        };
-
-        Ok(len)
-    }
-
-    pub async fn recv(&mut self) -> Option<Result<Row, Box<EvalAltResult>>> {
-        if self.done {
-            return None;
-        }
-
-        loop {
-            loop {
-                match self.codec.decode(&mut self.buf) {
-                    Ok(Some(line)) => match serde_json::from_str(&line) {
-                        Ok(res) => match res {
-                            QueryEvent::Columns(cols) => self.columns = Some(Arc::new(cols)),
-                            QueryEvent::Row { rowid, cells, .. } => match self.columns.as_ref() {
-                                Some(columns) => {
-                                    return Some(Ok(Row {
-                                        id: rowid,
-                                        columns: columns.clone(),
-                                        cells,
-                                    }));
-                                }
-                                None => {
-                                    self.done = true;
-                                    return Some(Err(Box::new(EvalAltResult::from(
-                                        "did not receive columns data",
-                                    ))));
-                                }
-                            },
-                            QueryEvent::EndOfQuery => {
-                                return None;
-                            }
-                            QueryEvent::Error(e) => {
-                                self.done = true;
-                                return Some(Err(Box::new(EvalAltResult::from(e))));
-                            }
-                        },
-                        Err(e) => {
-                            self.done = true;
-                            return Some(Err(Box::new(EvalAltResult::from(e.to_string()))));
-                        }
-                    },
-                    Ok(None) => {
-                        break;
-                    }
-                    Err(e) => {
-                        self.done = true;
-                        return Some(Err(Box::new(EvalAltResult::from(e.to_string()))));
-                    }
-                }
-            }
-            match self.read_bytes().await {
-                Ok(n) => {
-                    trace!("read {n} bytes");
-                }
-                Err(e) => return Some(Err(e)),
-            }
-        }
-    }
-}
-
-impl Iterator for QueryResponseIter {
-    type Item = Result<Row, Box<EvalAltResult>>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.done {
-            return None;
-        }
-        self.handle.clone().block_on(self.recv())
-    }
-}
-
-pub struct QueryHandle {
-    id: Uuid,
-    body: Option<hyper::Body>,
-    client: CorrosionApiClient,
-}
-
-impl QueryHandle {
-    async fn body(&mut self) -> Result<(hyper::Body, bool), corro_client::Error> {
-        if let Some(body) = self.body.take() {
-            return Ok((body, true));
-        }
-
-        Ok((self.client.watched_query(self.id).await?, false))
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum TemplateCommand {
-    Render,
-}
-
-pub trait WriteSeek: Write + Seek + Send + Sync + 'static {}
-impl<T> WriteSeek for T where T: Write + Seek + Send + Sync + 'static {}
-
-#[derive(Clone)]
-pub struct TemplateWriter(Arc<TemplateWriterInner>);
-
-pub struct TemplateWriterInner {
-    w: RwLock<Box<dyn WriteSeek>>,
-    tx: mpsc::Sender<TemplateCommand>,
-    cancel: CancellationToken,
-}
-
-impl TemplateWriter {
-    pub fn new<W: WriteSeek>(w: W, tx: mpsc::Sender<TemplateCommand>) -> Self {
-        let cancel = CancellationToken::new();
-        Self(Arc::new(TemplateWriterInner {
-            w: RwLock::new(Box::new(BufWriter::new(w))),
-            tx,
-            cancel,
-        }))
-    }
-
-    pub fn cancel(&self) {
-        self.0.cancel.cancel();
     }
 
     fn write_str(&mut self, data: &str) -> Result<(), Box<EvalAltResult>> {
@@ -407,9 +67,6 @@ impl TemplateWriter {
             return Ok(());
         }
         Ok(self
-            .0
-            .w
-            .write()
             .write_all(data.as_bytes())
             .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?)
     }
@@ -418,235 +75,45 @@ impl TemplateWriter {
         let mut b = [0; 2];
         let len = data.encode_utf8(&mut b).len();
         Ok(self
-            .0
-            .w
-            .write()
             .write_all(&b[0..len])
             .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?)
     }
 
     fn write_dynamic(&mut self, d: Dynamic) -> Result<(), Box<EvalAltResult>> {
         Ok(self
-            .0
-            .w
-            .write()
             .write_all(d.to_string().as_bytes())
             .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?)
     }
+}
 
-    fn write_sql_to_csv(&mut self, csv: SqlToCsv) -> Result<(), Box<EvalAltResult>> {
-        let mut rows = csv.res.into_iter();
+pub struct Engine {
+    engine: rhai::Engine,
+}
 
-        let mut w = self.0.w.write();
-        let mut wtr = csv::Writer::from_writer(&mut *w);
+impl Deref for Engine {
+    type Target = rhai::Engine;
 
-        let mut wrote_header = false;
-
-        while let Some(row) = rows.next() {
-            let row = row?;
-            if !wrote_header {
-                wtr.write_record(row.columns.as_slice())
-                    .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-                wrote_header = true;
-            }
-            wtr.serialize(row.cells.as_slice())
-                .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-        }
-        if !wrote_header {
-            if let Some(cols) = rows.columns.as_ref() {
-                wtr.write_record(cols.as_slice())
-                    .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn write_sql_to_json(&mut self, mut json: SqlToJson) -> Result<(), Box<EvalAltResult>> {
-        debug!("write_sql_to_json");
-        let mut rows = json.res.into_iter();
-        {
-            let mut w = self.0.w.write();
-            json.json_output.write_rows(&mut *w, &mut rows)?;
-        }
-
-        let tx = self.0.tx.clone();
-
-        let cancel = self.0.cancel.clone();
-
-        tokio::spawn(async move {
-            let row_recv = tokio::select! {
-                row_recv = rows.recv() => row_recv,
-                _ = cancel.cancelled() => {
-                    debug!("template cancellation trigger, returning from tokio task");
-                    return
-                },
-            };
-
-            match row_recv {
-                Some(Ok(row)) => {
-                    trace!("got an updated row! {:?}", row.cells);
-
-                    const DEBOUNCE_DEADLINE: Duration = Duration::from_millis(100);
-
-                    let deadline = sleep(DEBOUNCE_DEADLINE);
-                    tokio::pin!(deadline);
-                    loop {
-                        tokio::select! {
-                            None = rows.recv() => {
-                                debug!("end of iterator");
-                                break;
-                            },
-                            _ = &mut deadline => {
-                                debug!("deadline reached");
-                                break;
-                            }
-                        }
-                    }
-
-                    if let Err(_e) = tx.send(TemplateCommand::Render).await {
-                        debug!("could not send back re-render command, channel must be closed!");
-                    }
-                    return;
-                }
-                Some(Err(e)) => {
-                    // TODO: need to re-render possibly...
-                    warn!("error from upstream, returning... {e}");
-                    return;
-                }
-                None => {
-                    debug!("sql stream is done");
-                    return;
-                }
-            }
-        });
-
-        Ok(())
+    fn deref(&self) -> &Self::Target {
+        &self.engine
     }
 }
 
-#[derive(Clone)]
-pub struct Engine {
-    engine: Arc<rhai::Engine>,
+impl DerefMut for Engine {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.engine
+    }
 }
 
 impl Engine {
-    pub fn new(client: corro_client::CorrosionApiClient) -> Self {
+    pub fn new<S: Clone + Send + Sync + 'static>() -> Self {
         let mut engine = rhai::Engine::new();
 
-        engine.register_type::<TemplateWriter>();
-        engine.register_fn("write", TemplateWriter::write_str);
-        engine.register_fn("write", TemplateWriter::write_char);
-        engine.register_fn("write", TemplateWriter::write_sql_to_json);
-        engine.register_fn("write", TemplateWriter::write_sql_to_csv);
-        engine.register_fn("write", TemplateWriter::write_dynamic);
+        engine.register_type::<TemplateWriter<S>>();
+        engine.register_fn("write", TemplateWriter::<S>::write_str);
+        engine.register_fn("write", TemplateWriter::<S>::write_char);
+        engine.register_fn("write", TemplateWriter::<S>::write_dynamic);
 
-        engine.register_type_with_name::<QueryResponse>("QueryResponse");
-        engine.register_iterator_result::<QueryResponse, Row>();
-        engine.register_fn("to_json", QueryResponse::to_json);
-        engine.register_fn("to_json", QueryResponse::to_json_w_options);
-        engine.register_fn("to_csv", QueryResponse::to_csv);
-
-        engine.register_type_with_name::<Row>("Row");
-        engine.register_iterator::<Row>();
-
-        engine.register_type_with_name::<Cell>("Cell");
-        engine.register_fn("name", Cell::name);
-        engine.register_fn("value", Cell::value);
-
-        engine.register_type_with_name::<SqliteValueWrap>("SqliteValue");
-        engine.register_fn("to_json", SqliteValueWrap::to_json);
-        engine.register_fn("to_string", SqliteValueWrap::to_json);
-
-        fn sql(
-            client: &CorrosionApiClient,
-            stmt: Statement,
-        ) -> Result<QueryResponse, Box<EvalAltResult>> {
-            debug!("sql function call {stmt:?}");
-            let (query_id, body) = tokio::runtime::Handle::current()
-                .block_on(client.watch(&stmt))
-                .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-
-            debug!("got res w/ id: {query_id}");
-
-            let handle = Arc::new(TokioRwLock::new(QueryHandle {
-                id: query_id,
-                body: Some(body),
-                client: client.clone(),
-            }));
-
-            Ok(QueryResponse { query: handle })
-        }
-
-        engine.register_fn("sql", {
-            let client = client.clone();
-            move |query: &str| -> Result<QueryResponse, Box<EvalAltResult>> {
-                sql(&client, Statement::Simple(query.into()))
-            }
-        });
-
-        fn dyn_to_sql(v: Dynamic) -> Result<SqliteValue, Box<EvalAltResult>> {
-            Ok(match v.type_name() {
-                "()" => SqliteValue::Null,
-                "i64" => SqliteValue::Integer(
-                    v.as_int()
-                        .map_err(|_e| Box::new(EvalAltResult::from("could not cast to i64")))?,
-                ),
-                "f64" => SqliteValue::Real(
-                    v.as_float()
-                        .map_err(|_e| Box::new(EvalAltResult::from("could not cast to f64")))?,
-                ),
-                "bool" => {
-                    if v.as_bool()
-                        .map_err(|_e| Box::new(EvalAltResult::from("could not cast to bool")))?
-                    {
-                        SqliteValue::Integer(1)
-                    } else {
-                        SqliteValue::Integer(0)
-                    }
-                }
-                "blob" => SqliteValue::Blob(
-                    v.into_blob()
-                        .map_err(|_e| Box::new(EvalAltResult::from("could not cast to blob")))?
-                        .into(),
-                ),
-                // convert everything else into a string, including a string
-                _ => SqliteValue::Text(v.to_compact_string()),
-            })
-        }
-
-        engine.register_fn("sql", {
-            let client = client.clone();
-            move |query: &str, params: Vec<Dynamic>| -> Result<QueryResponse, Box<EvalAltResult>> {
-                let params = params
-                    .into_iter()
-                    .map(|v| dyn_to_sql(v))
-                    .collect::<Result<Vec<_>, _>>()?;
-                sql(&client, Statement::WithParams(query.into(), params))
-            }
-        });
-
-        engine.register_fn(
-            "sql",
-            move |query: &str, params: Map| -> Result<QueryResponse, Box<EvalAltResult>> {
-                let params = params
-                    .into_iter()
-                    .map(|(k, v)| Ok::<_, Box<EvalAltResult>>((k.to_string(), dyn_to_sql(v)?)))
-                    .collect::<Result<HashMap<_, _>, _>>()?;
-                sql(&client, Statement::WithNamedParams(query.into(), params))
-            },
-        );
-
-        engine.register_fn("hostname", || -> Result<String, Box<EvalAltResult>> {
-            Ok(hostname::get()
-                .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?
-                .to_string_lossy()
-                .to_string())
-        });
-
-        Self {
-            engine: Arc::new(engine),
-        }
+        Self { engine }
     }
 
     pub fn compile(&self, input: &str) -> Result<Template, CompileError> {
@@ -724,94 +191,26 @@ fn rhai_enquote(program: &mut String, text: &str, strip_newline: bool) {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-struct JsonOutput {
-    pretty: bool,
-    row_values_as_array: bool,
-}
-
-impl JsonOutput {
-    fn write_rows<W: Write>(
-        &mut self,
-        w: &mut W,
-        rows: &mut QueryResponseIter,
-    ) -> Result<(), Box<EvalAltResult>> {
-        if self.pretty {
-            let ser = serde_json::Serializer::pretty(w);
-            if self.row_values_as_array {
-                write_json_rows_as_array(ser, rows)
-            } else {
-                write_json_rows_as_object(ser, rows)
-            }
-        } else {
-            let ser = serde_json::Serializer::new(w);
-            if self.row_values_as_array {
-                write_json_rows_as_array(ser, rows)
-            } else {
-                write_json_rows_as_object(ser, rows)
-            }
-        }
-    }
-}
-
-fn write_json_rows_as_object<W: Write, F: Formatter>(
-    mut ser: serde_json::Serializer<W, F>,
-    rows: &mut QueryResponseIter,
-) -> Result<(), Box<EvalAltResult>> {
-    let mut seq = ser
-        .serialize_seq(None)
-        .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-    while let Some(row_res) = rows.next() {
-        let row = row_res.map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-
-        // we have to collect here due to serde limitations (I think), but it's not a big deal...
-        let map = row
-            .columns
-            .iter()
-            .enumerate()
-            .filter_map(|(i, col)| row.cells.get(i).map(|value| (col, value)))
-            .collect::<IndexMap<&CompactString, &SqliteValue>>();
-
-        seq.serialize_element(&map)
-            .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-    }
-    seq.end()
-        .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))
-}
-
-fn write_json_rows_as_array<W: Write, F: Formatter>(
-    mut ser: serde_json::Serializer<W, F>,
-    rows: &mut QueryResponseIter,
-) -> Result<(), Box<EvalAltResult>> {
-    let mut seq = ser
-        .serialize_seq(None)
-        .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-
-    while let Some(row_res) = rows.next() {
-        let row = row_res.map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-        seq.serialize_element(&row.cells)
-            .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
-    }
-
-    SerializeSeq::end(seq).map_err(|e| Box::new(EvalAltResult::from(e.to_string())))
-}
-
 pub struct Template<'a> {
     ast: rhai::AST,
     engine: &'a Engine,
 }
 
 impl<'a> Template<'a> {
-    pub fn render(&self, w: TemplateWriter) -> Result<(), Box<rhai::EvalAltResult>> {
+    pub fn render<W: Write + Send + Sync + 'static, S: Clone + Send + Sync + 'static>(
+        &self,
+        w: W,
+        state: S,
+    ) -> Result<(), Box<rhai::EvalAltResult>> {
         let mut scope = Scope::new();
+        let mut w = TemplateWriter::new(w, state);
         scope.push("__tpl_writer", w.clone());
+
         self.engine
             .engine
             .eval_ast_with_scope(&mut scope, &self.ast)?;
 
-        w.0.w
-            .write()
-            .flush()
+        w.flush()
             .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
 
         Ok(())
@@ -832,17 +231,10 @@ pub enum CompileError {
 
 #[cfg(test)]
 mod tests {
-    use corro_tests::launch_test_agent;
-    use tokio::{sync::mpsc, task::block_in_place};
-    use tripwire::Tripwire;
-
     use super::*;
 
-    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-    async fn test_basic() {
-        _ = tracing_subscriber::fmt::try_init();
-        let (tripwire, _trip_worker, _trip_sender) = Tripwire::new_simple();
-
+    #[test]
+    fn test_basic() {
         let tmpdir = tempfile::tempdir().unwrap();
         let filepath = tmpdir.path().join("output");
 
@@ -852,13 +244,7 @@ mod tests {
             .open(&filepath)
             .unwrap();
 
-        let ta = launch_test_agent(|conf| conf.build(), tripwire.clone())
-            .await
-            .unwrap();
-
-        let client = corro_client::CorrosionApiClient::new(ta.agent.api_addr());
-
-        let engine = Engine::new(client);
+        let engine = Engine::new::<()>();
 
         let input = r#"<%
 let a = [42, 123, 999, 0, true, "hello", "world!", 987.6543];
@@ -868,21 +254,16 @@ for (item, count) in a { %>
 Item #<%= count + 1 %> = <%= item %>
 <% } %>
 
-<%= hostname() %>
-
 tail"#;
 
-        let (tx, _rx) = mpsc::channel(1);
-
         let tpl = engine.compile(input).unwrap();
-        tpl.render(TemplateWriter::new(f, tx)).unwrap();
+        tpl.render(f, ()).unwrap();
 
         let output = std::fs::read_to_string(filepath).unwrap();
 
         assert_eq!(
             output,
-            format!(
-                "Item #1 = 42
+            "Item #1 = 42
 Item #2 = 123
 Item #3 = 999
 Item #4 = 0
@@ -891,128 +272,38 @@ Item #6 = hello
 Item #7 = world!
 Item #8 = 987.6543
 
-{}
-
 tail",
-                hostname::get().unwrap().to_string_lossy()
-            )
         );
     }
 
-    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-    async fn test_sql() {
-        _ = tracing_subscriber::fmt::try_init();
-        let (tripwire, _trip_worker, _trip_sender) = Tripwire::new_simple();
-
-        let ta = launch_test_agent(|conf| conf.build(), tripwire.clone())
-            .await
-            .unwrap();
-
-        let client = corro_client::CorrosionApiClient::new(ta.agent.api_addr());
-
-        client
-            .schema(&vec![Statement::Simple(corro_tests::TEST_SCHEMA.into())])
-            .await
-            .unwrap();
-
-        client
-            .execute(&vec![
-                Statement::WithParams(
-                    "insert into tests (id, text) values (?,?)".into(),
-                    vec!["service-id".into(), "service-name".into()],
-                ),
-                Statement::WithParams(
-                    "insert into tests (id, text) values (?,?)".into(),
-                    vec!["service-id-2".into(), "service-name-2".into()],
-                ),
-            ])
-            .await
-            .unwrap();
-
+    #[test]
+    fn test_extend() {
         let tmpdir = tempfile::tempdir().unwrap();
         let filepath = tmpdir.path().join("output");
 
-        // enum Command {
-        //     Render,
-        // }
-
-        // enum Reply {
-        //     Rendered,
-        // }
-
-        // let (cmd_tx, mut cmd_rx) = mpsc::channel(10);
-        // let (reply_tx, mut reply_rx) = mpsc::channel(10);
-
-        // tokio::spawn({
-        //     let filepath = filepath.clone();
-        //     async move {
         let f = std::fs::OpenOptions::new()
             .create(true)
             .write(true)
             .open(&filepath)
             .unwrap();
 
-        let engine = Engine::new(client.clone());
+        let mut engine = Engine::new::<()>();
 
-        // loop {
-        // let cmd = cmd_rx.recv().await.unwrap();
-        // assert!(matches!(cmd, Command::Render));
+        engine.register_fn(
+            "write",
+            |tw: &mut TemplateWriter<()>, d: Dynamic| -> Result<(), Box<EvalAltResult>> {
+                tw.write_all(format!("overloaded: {d}").as_bytes())
+                    .map_err(|e| Box::new(EvalAltResult::from(e.to_string())))?;
+                Ok(())
+            },
+        );
 
-        let (tx, mut rx) = mpsc::channel(1);
+        let tpl = engine.compile("<%= 123 %>").unwrap();
 
-        {
-            block_in_place(|| {
-                let input = r#"<%= sql("select * from tests").to_json(#{pretty: true}) %>
-<%= sql("select * from tests").to_json() %>"#;
+        tpl.render(f, ()).unwrap();
 
-                let tpl = engine.compile(input).unwrap();
-                tpl.render(TemplateWriter::new(f, tx)).unwrap();
-            });
-            // reply_tx.send(Reply::Rendered).await.unwrap();
-            // }
-            // }
-            // });
+        let output = std::fs::read_to_string(filepath).unwrap();
 
-            // cmd_tx.send(Command::Render).await.unwrap();
-            // let replied = reply_rx.recv().await.unwrap();
-            // assert!(matches!(replied, Reply::Rendered));
-
-            let output = std::fs::read_to_string(&filepath).unwrap();
-
-            println!("output: {output}");
-
-            client
-                .execute(&vec![Statement::WithParams(
-                    "insert into tests (id, text) values (?,?)".into(),
-                    vec!["service-id-3".into(), "service-name-3".into()],
-                )])
-                .await
-                .unwrap();
-
-            // we have 2 queries here...
-            assert_eq!(rx.recv().await.unwrap(), TemplateCommand::Render);
-            assert_eq!(rx.recv().await.unwrap(), TemplateCommand::Render);
-            assert!(rx.recv().await.is_none());
-        }
-
-        let f = std::fs::OpenOptions::new()
-            .create(true)
-            .write(true)
-            .open(&filepath)
-            .unwrap();
-
-        let (tx, _rx) = mpsc::channel(1);
-
-        block_in_place(|| {
-            let input = r#"<%= sql("select * from tests").to_json(#{pretty: true}) %>
-<%= sql("select * from tests").to_json() %>"#;
-
-            let tpl = engine.compile(input).unwrap();
-            tpl.render(TemplateWriter::new(f, tx)).unwrap();
-        });
-
-        let output = std::fs::read_to_string(&filepath).unwrap();
-
-        println!("output: {output}");
+        assert_eq!(output, "overloaded: 123");
     }
 }
